requirements.txt
Django
python-dotenv
psycopg2-binary
langchain-groq
requests
numpy
scikit-learn
pandas
pillow 
joblib
tensorflow
xgboost
fastai
